Revert "Add,delete and create notes functionality added and refactored code into separate request handlers and auth files"
