+++ tsconfig.json
@@ -1,11 +1,11 @@
 {
-  "compilerOptions": {
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext",
-    "target": "ES2020",
-    "sourceMap": true,
-    "outDir": "./dist/js/",
-    "esModuleInterop": true,
-  },
-  "include": ["src/**/*"],
-}
+    "compilerOptions": {
+      "module": "NodeNext",
+      "moduleResolution": "NodeNext",
+      "target": "ES2020",
+      "sourceMap": true,
+      "outDir": "./dist/js/",
+      "esModuleInterop": true,
+    },
+    "include": ["src/**/*"],
+  }
